scripts/requirements.txt
# Asset Pipeline Dependencies
Pillow>=10.0.0
requests>=2.31.0
toml>=0.10.2
click>=8.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional AI providers
# openai>=1.0.0
# replicate>=0.15.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0